backend/restful/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "yujhenchen",email = "yujhen.chen93@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg[binary] (>=3.2.5,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "djangorestframework-simplejwt[crypto] (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
